+++ tests/test_main.py
@@ -6,14 +6,14 @@ import pytest
 from pytest_mock import MockerFixture
 
 from lib import main
-from lib.config import Config
+from lib.config import AccountConfig, GlobalConfig, ReservationConfig
 from lib.notification_handler import NotificationHandler
 
 
 # We don't actually want the config to read the file for these tests
 @pytest.fixture(autouse=True)
 def mock_config(mocker: MockerFixture) -> None:
-    mocker.patch("lib.config.Config._read_config")
+    mocker.patch("lib.config.GlobalConfig._read_config")
 
 
 def test_print_version_prints_script_version(capsys: pytest.CaptureFixture[str]) -> None:
@@ -62,9 +62,31 @@ def test_check_flags_does_not_exit_when_flags_are_not_matched(
     mock_exit.assert_not_called()
 
 
+def test_get_notification_urls_gets_all_urls(mocker: MockerFixture) -> None:
+    config = GlobalConfig()
+    config.accounts = [AccountConfig()]
+    config.reservations = [ReservationConfig()]
+    config.notification_urls = ["url1"]
+    config.accounts[0].notification_urls = ["url1", "url2"]
+    config.reservations[0].notification_urls = ["url1", "url3"]
+
+    notification_urls = main.get_notification_urls(config)
+
+    # Sort because order is not important
+    assert sorted(notification_urls) == ["url1", "url2", "url3"]
+
+
+def test_test_notifications_sends_to_every_url_in_config(mocker: MockerFixture) -> None:
+    mock_send_notification = mocker.patch.object(NotificationHandler, "send_notification")
+
+    config = GlobalConfig()
+    main.test_notifications(config)
+    mock_send_notification.assert_called_once()
+
+
 def test_set_up_accounts_starts_all_accounts_in_proceses(mocker: MockerFixture) -> None:
-    config = Config()
-    config.accounts = [["user1", "pass1"], ["user2", "pass2"]]
+    config = GlobalConfig()
+    config.accounts = [AccountConfig(), AccountConfig()]
 
     mock_process = mocker.patch("lib.main.Process")
     mock_process.start = mock.Mock()
@@ -76,8 +98,8 @@ def test_set_up_accounts_starts_all_accounts_in_proceses(mocker: MockerFixture) 
 
 
 def test_set_up_reservations_starts_all_reservations_in_proceses(mocker: MockerFixture) -> None:
-    config = Config()
-    config.reservations = [["test1", "first1", "last1"], ["test2", "first2", "last2"]]
+    config = GlobalConfig()
+    config.reservations = [ReservationConfig(), ReservationConfig()]
 
     mock_process = mocker.patch("lib.main.Process")
     mock_process.start = mock.Mock()
@@ -88,13 +110,11 @@ def test_set_up_reservations_starts_all_reservations_in_proceses(mocker: MockerF
     assert mock_process.return_value.start.call_count == len(config.reservations)
 
 
-def test_set_up_check_in_sends_test_notifications_when_flag_is_passed(
-    mocker: MockerFixture,
-) -> None:
-    mock_send_notification = mocker.patch.object(NotificationHandler, "send_notification")
+def test_set_up_check_in_sends_test_notifications_when_flag_passed(mocker: MockerFixture) -> None:
+    mock_test_notifications = mocker.patch("lib.main.test_notifications")
     with pytest.raises(SystemExit):
         main.set_up_check_in(["--test-notifications"])
-    mock_send_notification.assert_called_once()
+    mock_test_notifications.assert_called_once()
 
 
 @pytest.mark.parametrize(